requirements.txt
Flask==3.1.0
openai==1.57.4
google-ai-generativelanguage==0.6.15
google-genai==1.0.0
google-generativeai==0.8.4
snowflake-connector-python==3.13.2
python-dotenv==1.0.1